requirements.txt
imgaug==0.4.0
numpy==1.18.2
pandas==1.0.3
torch==1.4.0
opencv_python==4.2.0.32
torchvision==0.5.0
pydicom==1.4.2
Pillow==7.1.2
scikit_learn==0.23.0
